svg/pservers/parsing/stop-opacity-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Paint Servers: parsing stop-opacity with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/pservers.html#StopOpacityProperty" rel="help"/><h:meta content="stop-opacity supports only the grammar 'alpha-value'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("stop-opacity", "1.");
test_invalid_value("stop-opacity", "2 3");

  ]]></script></svg>
